swms_backend/requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-extensions==3.0.8
django-filter==2.3.0
django-grappelli==2.14.2
django-guardian==2.3.0
django-import-export==2.3.0
django-model-utils==4.0.0
django-monthfield==0.1.3
django-pandas==0.6.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-yasg==1.17.1
gunicorn==20.0.4
Markdown==3.2.2
Pillow==7.2.0
psycopg2==2.8.6
Pygments==2.6.1
pylint==2.6.0
python-dotenv==0.14.0
rope==0.17.0
Werkzeug==1.0.1
whitenoise==5.2.0
PrettyTable==0.7.2
scikit-learn==0.23.2
deap==1.3.1
ipython==7.18.1
ipykernel==5.3.4
seaborn==0.10.1
